Rename Http/* first letter to uppercase
