+++ app/Http/Controllers/Dashboard/CollegeYearThreadController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\dashboard;
+namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Models\CollegeYearThread;
