+++ app/Http/Controllers/Dashboard/NotLoggedIn/DepartmentThreadController.php
@@ -1,14 +1,17 @@
 <?php
 
-namespace App\Http\Controllers\dashboard\not_logged_in;
+namespace App\Http\Controllers\Dashboard\NotLoggedIn;
 
 use App\Http\Controllers\Controller;
-use App\Models\LectureThread;
+use App\Models\DepartmentThread;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
-class LectureThreadController extends Controller
+class DepartmentThreadController extends Controller
 {
+    /** @var string */
+    private $thread_id;
+
     /**
      * Display a listing of the resource.
      *
@@ -52,8 +55,8 @@ class LectureThreadController extends Controller
     {
         $this->thread_id = $thread_id;
 
-        return LectureThread::select(
-            'lecture_threads.*',
+        return DepartmentThread::select(
+            'department_threads.*',
             DB::raw('COUNT(likes1.user_email) AS count_user'),
             DB::raw('0 AS user_like'),
             'thread_image_paths.img_path'
@@ -61,26 +64,26 @@ class LectureThreadController extends Controller
             ->leftjoin('likes AS likes1', function ($join) {
                 $join
                     ->where('likes1.thread_id', '=', $this->thread_id)
-                    ->whereColumn('likes1.message_id', '=', 'lecture_threads.message_id');
+                    ->whereColumn('likes1.message_id', '=', 'department_threads.message_id');
             })
             ->leftjoin('thread_image_paths', function ($join) {
                 $join
-                    ->whereColumn('thread_image_paths.thread_id', '=', 'lecture_threads.thread_id')
-                    ->whereColumn('thread_image_paths.message_id', '=', 'lecture_threads.message_id');
+                    ->whereColumn('thread_image_paths.thread_id', '=', 'department_threads.thread_id')
+                    ->whereColumn('thread_image_paths.message_id', '=', 'department_threads.message_id');
             })
-            ->where('lecture_threads.thread_id', '=', $this->thread_id)
-            ->where('lecture_threads.message_id', '>', $pre_max_message_id)
-            ->groupBy('lecture_threads.message_id')
+            ->where('department_threads.thread_id', '=', $this->thread_id)
+            ->where('department_threads.message_id', '>', $pre_max_message_id)
+            ->groupBy('department_threads.message_id')
             ->get();
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\LectureThread  $lectureThreads
+     * @param  \App\Models\DepartmentThread  $departmentThreads
      * @return \Illuminate\Http\Response
      */
-    public function edit(LectureThread $lectureThreads)
+    public function edit(DepartmentThread $departmentThreads)
     {
         //
     }
@@ -89,10 +92,10 @@ class LectureThreadController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\LectureThread  $lectureThreads
+     * @param  \App\Models\DepartmentThread  $departmentThreads
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, LectureThread $lectureThreads)
+    public function update(Request $request, DepartmentThread $departmentThreads)
     {
         //
     }
@@ -100,10 +103,10 @@ class LectureThreadController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\LectureThread  $lectureThreads
+     * @param  \App\Models\DepartmentThread  $departmentThreads
      * @return \Illuminate\Http\Response
      */
-    public function destroy(LectureThread $lectureThreads)
+    public function destroy(DepartmentThread $departmentThreads)
     {
         //
     }
