+++ app/Http/Controllers/Dashboard/NotLoggedIn/LectureThreadController.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Http\Controllers\dashboard\not_logged_in;
+namespace App\Http\Controllers\Dashboard\NotLoggedIn;
 
 use App\Http\Controllers\Controller;
-use App\Models\CollegeYearThread;
+use App\Models\LectureThread;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
-class CollegeYearThreadController extends Controller
+class LectureThreadController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -52,8 +52,8 @@ class CollegeYearThreadController extends Controller
     {
         $this->thread_id = $thread_id;
 
-        return CollegeYearThread::select(
-            'college_year_threads.*',
+        return LectureThread::select(
+            'lecture_threads.*',
             DB::raw('COUNT(likes1.user_email) AS count_user'),
             DB::raw('0 AS user_like'),
             'thread_image_paths.img_path'
@@ -61,26 +61,26 @@ class CollegeYearThreadController extends Controller
             ->leftjoin('likes AS likes1', function ($join) {
                 $join
                     ->where('likes1.thread_id', '=', $this->thread_id)
-                    ->whereColumn('likes1.message_id', '=', 'college_year_threads.message_id');
+                    ->whereColumn('likes1.message_id', '=', 'lecture_threads.message_id');
             })
             ->leftjoin('thread_image_paths', function ($join) {
                 $join
-                    ->whereColumn('thread_image_paths.thread_id', '=', 'college_year_threads.thread_id')
-                    ->whereColumn('thread_image_paths.message_id', '=', 'college_year_threads.message_id');
+                    ->whereColumn('thread_image_paths.thread_id', '=', 'lecture_threads.thread_id')
+                    ->whereColumn('thread_image_paths.message_id', '=', 'lecture_threads.message_id');
             })
-            ->where('college_year_threads.thread_id', '=', $this->thread_id)
-            ->where('college_year_threads.message_id', '>', $pre_max_message_id)
-            ->groupBy('college_year_threads.message_id')
+            ->where('lecture_threads.thread_id', '=', $this->thread_id)
+            ->where('lecture_threads.message_id', '>', $pre_max_message_id)
+            ->groupBy('lecture_threads.message_id')
             ->get();
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\CollegeYearThread  $collegeYearThreads
+     * @param  \App\Models\LectureThread  $lectureThreads
      * @return \Illuminate\Http\Response
      */
-    public function edit(CollegeYearThread $collegeYearThreads)
+    public function edit(LectureThread $lectureThreads)
     {
         //
     }
@@ -89,10 +89,10 @@ class CollegeYearThreadController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\CollegeYearThread  $collegeYearThreads
+     * @param  \App\Models\LectureThread  $lectureThreads
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, CollegeYearThread $collegeYearThreads)
+    public function update(Request $request, LectureThread $lectureThreads)
     {
         //
     }
@@ -100,10 +100,10 @@ class CollegeYearThreadController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\CollegeYearThread  $collegeYearThreads
+     * @param  \App\Models\LectureThread  $lectureThreads
      * @return \Illuminate\Http\Response
      */
-    public function destroy(CollegeYearThread $collegeYearThreads)
+    public function destroy(LectureThread $lectureThreads)
     {
         //
     }
