+++ app/Http/Controllers/MyPage/MyPageController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\mypage;
+namespace App\Http\Controllers\MyPage;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
