+++ routes/web.php
@@ -21,16 +21,16 @@ Route::middleware([
     Route::get('/', function () {
         return view('welcome');
     });
-    Route::get('dashboard', 'App\Http\Controllers\dashboard\DashboardController@threads')->name('dashboard');
-    Route::get('Q_and_A', 'App\Http\Controllers\Q_and_A\Q_and_AController@Q_and_A')->name('Q_and_A');
+    Route::get('dashboard', 'App\Http\Controllers\Dashboard\DashboardController@threads')->name('dashboard');
+    Route::get('Q_and_A', 'App\Http\Controllers\QandA\QandAController@Q_and_A')->name('Q_and_A');
 
     Route::middleware([
         'Access_log'
     ])->group(function () {
-        Route::get('dashboard/thread/name={thread_name}&id={thread_id}', 'App\Http\Controllers\dashboard\DashboardController@messages');
-        Route::get('dashboard/thread/name={thread_name}&id=', 'App\Http\Controllers\dashboard\DashboardController@messages');
-        Route::get('dashboard/thread/name=&id={thread_id}', 'App\Http\Controllers\dashboard\DashboardController@messages');
-        Route::get('dashboard/thread/name=&id=', 'App\Http\Controllers\dashboard\DashboardController@messages');
+        Route::get('dashboard/thread/name={thread_name}&id={thread_id}', 'App\Http\Controllers\Dashboard\DashboardController@messages');
+        Route::get('dashboard/thread/name={thread_name}&id=', 'App\Http\Controllers\Dashboard\DashboardController@messages');
+        Route::get('dashboard/thread/name=&id={thread_id}', 'App\Http\Controllers\Dashboard\DashboardController@messages');
+        Route::get('dashboard/thread/name=&id=', 'App\Http\Controllers\Dashboard\DashboardController@messages');
     });
 });
 
@@ -40,8 +40,8 @@ Route::middleware([
     config('jetstream.auth_session'),
     'verified'
 ])->group(function () {
-    Route::get('/mypage', 'App\Http\Controllers\mypage\MyPageController')->name('mypage');
-    Route::get('/report/create', 'App\Http\Controllers\report\FormContactAdministratorController@create')->name('report.create');
+    Route::get('/mypage', 'App\Http\Controllers\MyPage\MyPageController')->name('mypage');
+    Route::get('/report/create', 'App\Http\Controllers\Report\FormContactAdministratorController@create')->name('report.create');
 });
 
 // データ処理
@@ -53,34 +53,34 @@ Route::middleware([
         config('jetstream.auth_session'),
         'verified'
     ])->group(function () {
-        Route::controller(\App\Http\Controllers\dashboard\ThreadsController::class)->group(function () {
+        Route::controller(\App\Http\Controllers\Dashboard\ThreadsController::class)->group(function () {
             Route::match(['get', 'post'], 'jQuery.ajax/sendRow', 'store');
         });
 
-        Route::controller(\App\Http\Controllers\dashboard\HubController::class)->group(function () {
+        Route::controller(\App\Http\Controllers\Dashboard\HubController::class)->group(function () {
             Route::match(['get', 'post'], 'jQuery.ajax/create_thread', 'store');
         });
 
-        Route::controller(\App\Http\Controllers\dashboard\LikeController::class)->group(function () {
+        Route::controller(\App\Http\Controllers\Dashboard\LikeController::class)->group(function () {
             Route::match(['get', 'post'], 'jQuery.ajax/like', 'store');
             Route::match(['get', 'post'], 'jQuery.ajax/unlike', 'destroy');
         });
 
-        Route::controller(\App\Http\Controllers\mypage\UserController::class)->group(function () {
+        Route::controller(\App\Http\Controllers\MyPage\UserController::class)->group(function () {
             Route::match(['get', 'post'], 'jQuery.ajax/page_thema', 'update');
         });
 
-        Route::controller(\App\Http\Controllers\report\FormContactAdministratorController::class)->group(function () {
+        Route::controller(\App\Http\Controllers\Report\FormContactAdministratorController::class)->group(function () {
             Route::match(['get', 'post'], 'report/store', 'store')->name('report.store');
         });
     });
 
-    Route::controller(\App\Http\Controllers\dashboard\ThreadsController::class)->group(function () {
+    Route::controller(\App\Http\Controllers\Dashboard\ThreadsController::class)->group(function () {
         Route::match(['get', 'post'], 'jQuery.ajax/getRow', 'show');
     });
 });
 
-Route::controller(\App\Http\Controllers\mail\UserController::class)->group(function () {
+Route::controller(\App\Http\Controllers\Mail\UserController::class)->group(function () {
     // アカウント登録キャンセル
     Route::get('/account/delete/{id}/{hash}', 'destroy')->middleware('auth')->name('account/delete');
 });
